Improved Leaderboard accuracy during race & circuit rotation
The leaderboard is now calculated by plotting the drivers' progress along the track instead of relying on the "Distance" and "RelativeDistance" data from the telemetry which can sometimes be inaccurate in relation to the XY position data that we see on-screen.

Also, the circuits are now rotated to their correct orientation that is seen when the circuits are shown on TV or online. :)